move all Minimal Pools to integration testing
