+++ src/Estimators/tests/test_EstimatorManagerCrowd.cpp
@@ -21,9 +21,9 @@
 #include "Estimators/tests/EstimatorManagerNewTest.h"
 #include "Estimators/EstimatorInputDelegates.h"
 #include "EstimatorManagerInputTest.h"
-#include "Particle/tests/MinimalParticlePool.h"
-#include "QMCHamiltonians/tests/MinimalHamiltonianPool.h"
-#include "QMCWaveFunctions/tests/MinimalWaveFunctionPool.h"
+#include <MinimalParticlePool.h>
+#include <MinimalHamiltonianPool.h>
+#include <MinimalWaveFunctionPool.h>
 #include "Utilities/ResourceCollection.h"
 #include "Utilities/ProjectData.h"
