+++ src/Particle/LongRange/tests/test_kcontainer.cpp
@@ -13,7 +13,7 @@
 
 #include "Particle/LongRange/KContainer.h"
 #include "Particle/Lattice/CrystalLattice.h"
-#include "Particle/tests/MinimalParticlePool.h"
+#include <MinimalParticlePool.h>
 #include "Utilities/for_testing/checkVector.hpp"
 #include "Utilities/for_testing/NativeInitializerPrint.hpp"
 
@@ -39,8 +39,7 @@ TEST_CASE("kcontainer at gamma in 3D", "[longrange]")
   for (int ik = 0; ik < kcs.size(); ik++)
   {
     const double kc = kcs[ik] + 1e-6;
-    klists.updateKLists(lattice, kc,
-                                                                                                         ndim);
+    klists.updateKLists(lattice, kc, ndim);
     CHECK(klists.getKpts().size() == nks[ik]);
   }
   const int mxk    = klists.getKpts().size();
@@ -76,14 +75,12 @@ TEST_CASE("kcontainer at twist in 3D", "[longrange]")
 
   PosType twist;
   twist[0] = 0.1;
-  klists.updateKLists(lattice, kc,
-                                                                                                       ndim, twist);
+  klists.updateKLists(lattice, kc, ndim, twist);
   CHECK(klists.getKpts().size() == 1);
   CHECK(klists.getKptsCartWorking()[0][0] == Approx(blat * (twist[0] - 1)));
 
   twist = {-0.5, 0, 0.5};
-  klists.updateKLists(lattice, kc,
-                                                                                                       ndim, twist);
+  klists.updateKLists(lattice, kc, ndim, twist);
   int gvecs[3][3] = {{0, 0, -1}, {1, 0, -1}, {1, 0, 0}};
   CHECK(klists.getKpts().size() == 3);
   for (int ik = 0; ik < klists.getKpts().size(); ik++)
@@ -109,8 +106,7 @@ TEST_CASE("kcontainer at gamma in 2D", "[longrange]")
   for (int ik = 0; ik < kcs.size(); ik++)
   {
     const double kc = kcs[ik] + 1e-6;
-    klists.updateKLists(lattice, kc,
-                                                                                                         ndim);
+    klists.updateKLists(lattice, kc, ndim);
     CHECK(klists.getKpts().size() == nks[ik]);
   }
   const int mxk    = klists.getKpts().size();
@@ -146,14 +142,12 @@ TEST_CASE("kcontainer at twist in 2D", "[longrange]")
 
   PosType twist;
   twist[0] = 0.1;
-  klists.updateKLists(lattice, kc,
-                                                                                                       ndim, twist);
+  klists.updateKLists(lattice, kc, ndim, twist);
   CHECK(klists.getKpts().size() == 1);
   CHECK(klists.getKptsCartWorking()[0][0] == Approx(blat * (twist[0] - 1)));
 
   twist[1] = 0.1;
-  klists.updateKLists(lattice, kc,
-                                                                                                       ndim, twist);
+  klists.updateKLists(lattice, kc, ndim, twist);
   CHECK(klists.getKpts().size() == 2);
   CHECK(klists.getKptsCartWorking()[0][0] == Approx(blat * (twist[0] - 1)));
   CHECK(klists.getKptsCartWorking()[0][1] == Approx(blat * twist[1]));
@@ -161,8 +155,7 @@ TEST_CASE("kcontainer at twist in 2D", "[longrange]")
   CHECK(klists.getKptsCartWorking()[1][1] == Approx(blat * (twist[1] - 1)));
 
   twist = {-0.5, 0.5, 0};
-  klists.updateKLists(lattice, kc,
-                                                                                                       ndim, twist);
+  klists.updateKLists(lattice, kc, ndim, twist);
   CHECK(klists.getKpts().size() == 3);
   //for (int ik=0;ik<3;ik++)
   //  app_log() << klists.getKptsCartWorking()[ik] << std::endl;
@@ -189,7 +182,7 @@ TEST_CASE("kcontainer for diamond", "[longrange]")
   lattice.LR_dim_cutoff = 15;
   SimulationCell cell(lattice);
 
-  const KContainer& klists = cell.getKLists();
+  const KContainer& klists                                                                 = cell.getKLists();
   std::remove_cv_t<std::remove_reference_t<decltype(KContainer().getKpts())>> kpoint_lists = {
       {-1, -1, -1}, {-1, 0, 0},   {0, -1, 0},   {0, 0, -1},   {0, 0, 1},    {0, 1, 0},    {1, 0, 0},    {1, 1, 1},
       {-1, -1, 0},  {-1, 0, -1},  {0, -1, -1},  {0, 1, 1},    {1, 0, 1},    {1, 1, 0},    {-2, -1, -1}, {-1, -2, -1},
