+++ src/QMCDrivers/tests/CMakeLists.txt
@@ -43,7 +43,7 @@ if(HAVE_MPI)
 endif()
 add_executable(${UTEST_EXE} ${DRIVER_TEST_SRC})
 
-target_link_libraries(${UTEST_EXE} catch_main qmcdriver)
+target_link_libraries(${UTEST_EXE} catch_main qmcdriver minimal_testing_pools)
 if(USE_OBJECT_TARGET)
   target_link_libraries(
     ${UTEST_EXE}
@@ -87,7 +87,7 @@ set(DRIVER_TEST_SRC
     test_Eigensolver.cpp
     test_WFOptDriverInput.cpp)
 add_executable(${UTEST_EXE} ${DRIVER_TEST_SRC})
-target_link_libraries(${UTEST_EXE} catch_main qmcdriver)
+target_link_libraries(${UTEST_EXE} catch_main qmcdriver minimal_testing_pools)
 if(USE_OBJECT_TARGET)
   target_link_libraries(
     ${UTEST_EXE}
@@ -111,7 +111,7 @@ if(HAVE_MPI)
   #this is dependent on the directory creation and sym linking of earlier driver tests
   set(DRIVER_TEST_SRC SetupPools.cpp test_WalkerControl.cpp test_QMCDriverNew.cpp)
   add_executable(${UTEST_EXE} ${DRIVER_TEST_SRC})
-  target_link_libraries(${UTEST_EXE} catch_main qmcdriver)
+  target_link_libraries(${UTEST_EXE} catch_main qmcdriver minimal_testing_pools)
   if(USE_OBJECT_TARGET)
     target_link_libraries(
       ${UTEST_EXE}
