+++ src/QMCDrivers/tests/test_WalkerLogCollector.cpp
@@ -14,9 +14,9 @@
 
 #include "Message/Communicate.h"
 #include "QMCHamiltonians/QMCHamiltonian.h"
-#include "Particle/tests/MinimalParticlePool.h"
-#include "QMCHamiltonians/tests/MinimalHamiltonianPool.h"
-#include "QMCWaveFunctions/tests/MinimalWaveFunctionPool.h"
+#include <MinimalParticlePool.h>
+#include <MinimalHamiltonianPool.h>
+#include <MinimalWaveFunctionPool.h>
 #include "Utilities/ResourceCollection.h"
 #include "Utilities/ProjectData.h"
