+++ src/integration_testing/MinimalHamiltonianPool.h
