+++ src/integration_testing/MinimalParticlePool.h
